% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definitions.R
\name{calculate_bias_variance}
\alias{calculate_bias_variance}
\title{Calculate Bias and Variance Components}
\usage{
calculate_bias_variance(all_predictions, test_data)
}
\arguments{
\item{all_predictions}{List of prediction matrices}

\item{test_data}{Test dataset}
}
\value{
Bias-variance decomposition
}
\description{
Calculate Bias and Variance Components
}
