% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definitions.R
\name{get_complexity_models}
\alias{get_complexity_models}
\title{Get Models with Different Complexity Levels}
\usage{
get_complexity_models()
}
\value{
Named list of model functions with increasing complexity
}
\description{
Get Models with Different Complexity Levels
}
