% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_model_recommendations}
\alias{get_model_recommendations}
\title{Get Model Recommendations}
\usage{
get_model_recommendations(x)
}
\arguments{
\item{x}{A bias_variance_analysis object}
}
\value{
A character vector of recommendations
}
\description{
Provides recommendations based on bias-variance analysis results
}
