% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_bias_variance_app}
\alias{launch_bias_variance_app}
\title{Launch Bias-Variance Analysis App}
\usage{
launch_bias_variance_app(...)
}
\arguments{
\item{...}{Arguments passed to run_app()}
}
\value{
Shiny app object
}
\description{
Alternative function name for launching the Shiny application
}
\examples{
if (interactive()) {
  launch_bias_variance_app()
}
}
