% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.bias_variance_analysis}
\alias{plot.bias_variance_analysis}
\title{Plot method for bias_variance_analysis objects}
\usage{
\method{plot}{bias_variance_analysis}(x, type = "bias_variance", ...)
}
\arguments{
\item{x}{A bias_variance_analysis object}

\item{type}{Character string: "bias_variance" or "performance"}

\item{...}{Additional arguments}
}
\description{
Plot method for bias_variance_analysis objects
}
