% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_bias_variance}
\alias{plot_bias_variance}
\title{Plot Bias-Variance Decomposition}
\usage{
plot_bias_variance(x, ...)
}
\arguments{
\item{x}{A bias_variance_analysis object}

\item{...}{Additional arguments passed to ggplot}
}
\value{
A ggplot object
}
\description{
Creates visualization of bias-variance tradeoff from analysis results
}
