% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_plotting.R
\name{plot_clinical_dashboard}
\alias{plot_clinical_dashboard}
\title{Create Clinical Performance Dashboard}
\usage{
plot_clinical_dashboard(metrics_results)
}
\arguments{
\item{metrics_results}{Results from simulation}
}
\value{
Plotly object
}
\description{
Create Clinical Performance Dashboard
}
