% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definitions.R
\name{run_bias_variance_analysis}
\alias{run_bias_variance_analysis}
\title{Run Bias-Variance Decomposition}
\usage{
run_bias_variance_analysis(n_bootstrap = 50, n_patients = 800, test_size = 200)
}
\arguments{
\item{n_bootstrap}{Number of bootstrap samples}

\item{n_patients}{Sample size per bootstrap}

\item{test_size}{Size of test set}
}
\value{
Bias-variance analysis results
}
\description{
Run Bias-Variance Decomposition
}
